fix(core): unexpected redirect to expired page after accepting invitation
